requirements.txt
django>=2.1
django-cors-headers>=3.0.2
djangorestframework>=3.9
djangorestframework-simplejwt>=4.3.0
psycopg2>=2.7
